% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstras.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algoritm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{as dataframe}

\item{init_node}{as numeric}
}
\value{
Shortest distance to all vertex from init_node.
}
\description{
Dijkstra Algoritm
}
\examples{
dijkstra(graph = wiki_graph,init_node = 1)


}
\references{
<https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm>
}
\author{
Maria,Masinde,Omkar
}
