% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Function to get greatest common divisor
Authors: Maria Treesa Sebastian(marse306), Brian Masinde(brima748), Omkar(omkbh878)}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{as numeric}

\item{y}{as numeric}
}
\value{
GCD of x and y
}
\description{
Function to get greatest common divisor
Authors: Maria Treesa Sebastian(marse306), Brian Masinde(brima748), Omkar(omkbh878)
}
\examples{
euclidean(100,10000)

}
