% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Weights attached to source vertices and destination vertices.}
\format{A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{id of the source vertices}
  \item{v2}{id of the destination vertices}
  \item{w}{weight of the path taken}
  ...
}}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the weights attached to traversing from sources to destinations.
}
\keyword{datasets}
