% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstras.R
\name{dijkstra}
\alias{dijkstra}
\title{Function to get the shortest distance to all the nodes from an initial node
Authors: Maria Treesa Sebastian(marse306), Brian Masinde(brima748), Omkar(omkbh878)}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{as dataframe}

\item{init_node}{as numeric}
}
\value{
Shortest distance to all vertex from init_node . The wikipage link to algorithm : <https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm>
}
\description{
Function to get the shortest distance to all the nodes from an initial node
Authors: Maria Treesa Sebastian(marse306), Brian Masinde(brima748), Omkar(omkbh878)
}
\examples{
dijkstra(graph = data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)),init_node = 1)


}
