% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Eucledian Algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{as numeric}

\item{y}{as numeric}
}
\value{
GCD of x and y .
}
\description{
Eucledian Algorithm
}
\examples{
euclidean(100,10000)

}
\references{
<https://en.wikipedia.org/wiki/Euclidean_algorithm>
}
\author{
Maria ,Masinde, Omkar
}
